+++ packages/core/src/lib/oauth/checks.ts
@@ -9,6 +9,7 @@ import type {
   RequestInternal,
 } from "../../types.js"
 import type { Cookie } from "../utils/cookie.js"
+import { OAuthConfigInternal } from "../../providers/oauth.js"
 
 interface CheckPayload {
   value: string
@@ -231,4 +232,33 @@ export const nonce = {
 
     return value.value
   },
+}
+
+/**
+ * When the authorization flow contains a state, we check if it's a redirect proxy
+ * and if so, we return the random state and the original redirect URL.
+ */
+export function handleState(
+  query: RequestInternal["query"],
+  provider: OAuthConfigInternal<any>,
+  isOnRedirectProxy: InternalOptions["isOnRedirectProxy"]
+) {
+  let randomState: string | undefined
+  let proxyRedirect: string | undefined
+
+  if (provider.redirectProxyUrl && !query?.state) {
+    throw new InvalidCheck(
+      "Missing state in query, but required for redirect proxy"
+    )
+  }
+
+  const state = decodeState(query?.state)
+  randomState = state?.random
+
+  if (isOnRedirectProxy) {
+    if (!state?.origin) return { randomState }
+    proxyRedirect = `${state.origin}?${new URLSearchParams(query)}`
+  }
+
+  return { randomState, proxyRedirect }
 }
